feat: add setup for daily interaction + minors changes
F01 - add a setup function called init to verify if the user is
interacting with the app for the first time in the day, and then give
the set of tasks to be performed. The same tasks will be given to the
user as long as 1- its not marked as complete; 2- was not deleted, 3-
its still the same day util the last interaction.

C01 - When a task is marked as complete it will be removed from the todo
tasks slice, and if the tasks is "repeteable" then it be reiserted in
all tasks slice, otherwise it will be thrown alway

C02 - Generates a json file called saves.json even when it doest exist